refactor(framework): move makeProps to utils
